--- app/controllers/concerns/authorization.rb
@@ -1,16 +1,0 @@
-module Authorization
-  extend ActiveSupport::Concern
-
-  included do
-    include Pundit
-
-    rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
-
-    private
-
-    def user_not_authorized
-      flash[:alert] = t('pundit.not_authorized')
-      redirect_to(request.referrer || root_path)
-    end
-  end
-end
